modification de l'orchestrator: gestion des MS

+++ atelier-1/api-common/src/main/java/fr/cpe/scoobygang/common/repository/ActiveMQTransactionRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import java.util.Optional;
 
-public interface ActiveMQTransactionRepository extends CrudRepository<ActiveMQTransaction, Integer>  {
+public interface ActiveMQTransactionRepository extends CrudRepository<ActiveMQTransaction, Long>  {
     Optional<ActiveMQTransaction> findByUuid(String uuid);
 }
